requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
bleach==5.0.0
blis==0.7.5
cachelib==0.4.1
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
cymem==2.0.6
debugpy==1.5.1
decorator==4.4.2
deepspeech==0.9.3
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
errant==2.3.3
et-xmlfile==1.1.0
fastjsonschema==2.15.3
filelock==3.0.12
Flask==2.0.2
Flask-Session==0.4.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.29.1
fsspec==2022.3.0
fuzzywuzzy==0.18.0
gast==0.5.3
gdown==4.4.0
gevent==21.12.0
google==3.0.0
google-auth==1.4.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gramformer @ git+https://github.com/PrithivirajDamodaran/Gramformer.git@41e396937a122e10294e04069ad89430f39fa778
greenlet==1.1.2
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.5.1
idna==2.8
imageio==2.16.2
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
inflect==5.3.0
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jarowinkler==1.0.2
jedi==0.18.1
Jinja2==3.0.2
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.4.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langcodes==3.3.0
libclang==13.0.0
librosa==0.8.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
moviepy==1.0.3
multiprocess==0.70.12.2
murmurhash==1.0.6
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.1
networkx==2.8
nltk==3.6.2
notebook==6.4.10
numba
numpy==1.20.3
oauthlib==3.2.0
omegaconf==2.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathlib==1.0.1
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2021.5.29
plac==1.1.3
pooch==1.5.1
portpicker==1.2.0
preshed==3.0.6
proglog==0.1.9
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydub==0.25.1
Pygments==2.10.0
pynndescent==0.5.4
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-engineio==4.3.1
python-Levenshtein==0.12.2
python-socketio==5.5.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
rapidfuzz==2.0.9
regex==2021.11.2
requests==2.21.0
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
sacremoses==0.0.49
scikit-image==0.18.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
sentencepiece==0.1.96
simplegeneric==0.8.1
six==1.12.0
smart-open==5.2.1
sounddevice==0.4.3
SoundFile==0.10.3.post1
soupsieve==2.2.1
spacy==2.3.7
spacy-legacy==3.0.8
spacy-loggers==1.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.35
srsly==1.0.5
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
thinc==7.4.5
threadpoolctl==2.2.0
tifffile==2022.4.8
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.8.1
torchaudio==0.8.1
torchfile==0.1.0
torchvision==0.2.1
tornado==6.1
tqdm==4.45.0
traitlets==5.1.1
transformers==4.18.0
typer==0.4.0
typing-extensions==3.10.0.2
umap-learn==0.5.2
Unidecode==1.3.2
urllib3==1.24.3
virtualenv==20.4.7
virtualenv-clone==0.5.4
visdom==0.1.8.9
wasabi==0.9.0
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.14.0
xlrd==2.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
